separate out __init__ file for metadata pkg
